refactor(chat): extract handle_chat_send from run_agent_turn
Phase 1 of the dual-mode chat refactor approved at
docs/superpowers/specs/2026-06-24-chat-bridge-architecture-design.md
and planned in ~/.claude/plans/i-have-an-idea-enchanted-flurry.md.

Move the body of jarvis.chat.worker.run_agent_turn and every helper
it owns (the _AssistantContentBatcher class, _resolve_model_and_
provider, _create_assistant_placeholder, _mark_errored, _handle_
event, _handle_event_inner, _prepend_doc_context, _vision_enabled,
_to_managed_attachments, _prepare_attachments, plus their module-
level constants) into a new module jarvis.chat.turn_handler with a
single public entry point: handle_chat_send(payload: dict) -> None.
The shape is caller-context-agnostic so the future Path B realtime-
process subscriber (gated on socketio_backend == 'python') can call
the same code without duplicating it.

jarvis.chat.worker keeps run_agent_turn as a thin shim that builds
the payload dict from the RQ positional args and forwards to
handle_chat_send, and re-exports every helper so existing tests
importing them via jarvis.chat.worker keep working unchanged.

Verification:
- bench --site jarvis-test.localhost run-tests --module
  jarvis.tests.test_chat_worker -> 13/15 pass, 2 fail. The two
  failing tests (test_prepends_today_context_to_user_message and
  test_tool_events_persist_tool_message_rows) are pre-existing
  failures on origin/main; confirmed by stashing the refactor and
  running the same suite on bare main, which produces the same two
  failures.
- bench --site jarvis-test.localhost run-tests --module
  jarvis.tests.test_turn_handler -> 4/4 pass.

No behavioural change. No SPA changes. No Procfile changes. No
config changes. Pure code organisation.
